@@ -1,86 +1,189 @@
-name: slt_app
-description: SLT Flutter application with reusable widgets and clean architecture.
+name: spaced_learning_app
+description: A Flutter app for spaced learning with the Spaced Learning API.
+# The following line prevents the package from being accidentally published to
+# pub.dev using `flutter pub publish`. This is preferred for private packages.
+publish_to: 'none' # Remove this line if you wish to publish to pub.dev
 
-# Prevent breaking changes for the future:
-publish_to: 'none'
-
-# Versions follow Semantic Versioning (https://semver.org/)
+# The following defines the version and build number for your application.
+# A version number is three numbers separated by dots, like 1.2.43
+# followed by an optional build number separated by a +.
+# Both the version and the builder number may be overridden in flutter
+# build by specifying --build-name and --build-number, respectively.
+# In Android, build-name is used as versionName while build-number used as versionCode.
+# Read more about Android versioning at https://developer.android.com/studio/publish/versioning
+# In iOS, build-name is used as CFBundleShortVersionString while build-number is used as CFBundleVersion.
+# Read more about iOS versioning at
+# https://developer.apple.com/library/archive/documentation/General/Reference/InfoPlistKeyReference/Articles/CoreFoundationKeys.html
+# In Windows, build-name is used as the major, minor, and patch parts
+# of the product and file versions while build-number is used as the build suffix.
 version: 1.0.0+1
 
 environment:
-  sdk: '>=3.0.0 <4.0.0'
+  sdk: ^3.8.0-231.0.dev
 
+# Dependencies specify other packages that your package needs in order to work.
+# To automatically upgrade your package dependencies to the latest versions
+# consider running `flutter pub upgrade --major-versions`. Alternatively,
+# dependencies can be manually updated by changing the version numbers below to
+# the latest version available on pub.dev. To see which dependencies have newer
+# versions available, run `flutter pub outdated`.
 dependencies:
   flutter:
     sdk: flutter
+  # The following adds the Cupertino Icons font to your application.
+  # Use with the CupertinoIcons class for iOS style icons.
+  cupertino_icons: ^1.0.8
+  go_router: ^15.1.1
 
-  # UI related
-  cupertino_icons: ^1.0.2
-  flutter_svg: ^2.0.5
-  cached_network_image: ^3.2.3
-  shimmer: ^3.0.0
-  flutter_staggered_animations: ^1.1.1
+  # Architecture & State Management
+  provider: ^6.0.5
+  get_it: ^8.0.3
+  event_bus: ^2.0.0
 
-  # State management
-  flutter_riverpod: ^2.3.6
-  riverpod_annotation: ^2.1.1
+  # Network & API
+  dio: ^5.3.2
+  json_annotation: ^4.9.0
+  http_interceptor: ^2.0.0
 
-  # Navigation
-  go_router: ^10.0.0
-
-  # Data handling
-  freezed_annotation: ^2.2.0
-  json_annotation: ^4.8.1
-  equatable: ^2.0.5
-
-  # Network
-  dio: ^5.2.1+1
-  connectivity_plus: ^4.0.1
-  http: ^1.1.0
-
-  # Storage
-  shared_preferences: ^2.1.2
-  flutter_secure_storage: ^8.0.0
-
-  # Utilities
-  get_it: ^7.6.0
-  logger: ^1.4.0
+  # Utils and UI enhancement
   intl: ^0.20.2
-  flutter_dotenv: ^5.1.0
+  shared_preferences: ^2.2.1
+  flutter_secure_storage: ^9.2.4
+  flutter_svg: ^2.0.7
+  freezed_annotation: ^3.0.0
+  fluttertoast: ^8.2.2
+  skeletonizer: ^1.4.3
 
-  # Localization
-  flutter_localizations:
-    sdk: flutter
+  # UI Components
+  google_fonts: ^6.2.1
+  flutter_spinkit: ^5.2.0
+  shimmer: ^3.0.0
+  flex_color_scheme: ^8.2.0
+
+  # Reminder & Notification
+  firebase_core: ^3.13.0
+  firebase_messaging: ^15.2.5
+  device_info_plus: ^11.3.3
+  permission_handler: ^12.0.0+1
+  flutter_local_notifications: ^19.0.0
+  android_alarm_manager_plus: ^4.0.7
+  timezone: ^0.10.0
+
+  #logo
+  flutter_launcher_icons: ^0.14.3
+  flutter_native_splash: ^2.3.10
+
+  # Riverpod
+  flutter_riverpod: ^2.6.1
+  riverpod_annotation: ^2.6.1
+  url_launcher: ^6.1.14
 
 dev_dependencies:
   flutter_test:
     sdk: flutter
+  mocktail: ^1.0.0 # giả lập dữ liệu cho unit test
+  integration_test:
+    sdk: flutter # nếu viết integration test
 
-  # Static analysis & linting
-  flutter_lints: ^2.0.1
+  # The "flutter_lints" package below contains a set of recommended lints to
+  # encourage good coding practices. The lint set provided by the package is
+  # activated in the `analysis_options.yaml` file located at the root of your
+  # package. See that file for information about deactivating specific lint
+  # rules and activating additional ones.
+  flutter_lints: ^5.0.0
+  build_runner: ^2.4.15
+  json_serializable: ^6.9.5
+  freezed: ^3.0.6
 
   # Code generation
-  build_runner: ^2.4.6
-  riverpod_generator: ^2.2.3
-  freezed: ^2.3.5
-  json_serializable: ^6.7.1
+  riverpod_generator: ^2.6.5
+  custom_lint: ^0.7.5
+  riverpod_lint: ^2.3.7
 
+# Cấu hình logo cho ứng dụng
+flutter_launcher_icons:
+  android: true
+  ios: true
+  image_path: "assets/images/app_icon.png"
+  min_sdk_android: 21
+  adaptive_icon_background: "#166683" # Primary color của FlexScheme.shark
+  adaptive_icon_foreground: "assets/images/app_icon.png" # Dùng chung một file
+  remove_alpha_ios: false
+  web:
+    generate: false
+  windows:
+    generate: false
+  macos:
+    generate: false
+  linux:
+    generate: false
+
+# Cấu hình splash screen
+flutter_native_splash:
+  color: "#166683" # Primary color của FlexScheme.shark
+  image: assets/images/app_icon.png
+  android_12:
+    image: assets/images/app_icon.png
+    color: "#166683"
+    icon_background_color: "#C0E8FF" # Primary container color
+
+  # Các tùy chọn khác
+  android: true
+  ios: true
+  web: false
+
+  # Tùy chọn hiển thị
+  fullscreen: true
+
+  # Tùy chọn nâng cao
+  android_gravity: center
+  ios_content_mode: center
+
+  # Thay đổi cấu hình thanh trạng thái - sử dụng tham số được hỗ trợ
+  # Loại bỏ: android_status_bar_color: "#166683"
+  # status_bar_color: "#166683" # Tham số đúng
+  # ios_status_bar_style: dark-content
+
+# For information on the generic Dart part of this file, see the
+# following page: https://dart.dev/tools/pub/pubspec
+
+# The following section is specific to Flutter packages.
 flutter:
-  uses-material-design: true
 
+  # The following line ensures that the Material Icons font is
+  # included with your application, so that you can use the icons in
+  # the material Icons class.
+  uses-material-design: true
   assets:
     - assets/images/
-    - assets/icons/
-    - assets/fonts/
-    - .env
 
-  fonts:
-    - family: Poppins
-      fonts:
-        - asset: assets/fonts/Poppins-Regular.ttf
-        - asset: assets/fonts/Poppins-Medium.ttf
-          weight: 500
-        - asset: assets/fonts/Poppins-SemiBold.ttf
-          weight: 600
-        - asset: assets/fonts/Poppins-Bold.ttf
-          weight: 700
+  # To add assets to your application, add an assets section, like this:
+  # assets:
+  #   - images/a_dot_burr.jpeg
+  #   - images/a_dot_ham.jpeg
+
+  # An image asset can refer to one or more resolution-specific "variants", see
+  # https://flutter.dev/to/resolution-aware-images
+
+  # For details regarding adding assets from package dependencies, see
+  # https://flutter.dev/to/asset-from-package
+
+  # To add custom fonts to your application, add a fonts section here,
+  # in this "flutter" section. Each entry in this list should have a
+  # "family" key with the font family name, and a "fonts" key with a
+  # list giving the asset and other descriptors for the font. For
+  # example:
+  # fonts:
+  #   - family: Schyler
+  #     fonts:
+  #       - asset: fonts/Schyler-Regular.ttf
+  #       - asset: fonts/Schyler-Italic.ttf
+  #         style: italic
+  #   - family: Trajan Pro
+  #     fonts:
+  #       - asset: fonts/TrajanPro.ttf
+  #       - asset: fonts/TrajanPro_Bold.ttf
+  #         weight: 700
+  #
+  # For details regarding fonts from package dependencies,
+  # see https://flutter.dev/to/font-from-package
