@@ -1,269 +1,64 @@
+// lib/presentation/screens/auth/login_screen.dart
 import 'package:flutter/material.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
-import 'package:slt_app/core/constants/app_assets.dart';
-import 'package:slt_app/core/di/dependency_injection.dart';
-import 'package:slt_app/core/router/app_routes.dart';
-import 'package:slt_app/presentation/widgets/buttons/slt_primary_button.dart';
-import 'package:slt_app/presentation/widgets/common/slt_app_bar.dart';
-import 'package:slt_app/presentation/widgets/inputs/slt_password_text_field.dart';
-import 'package:slt_app/presentation/widgets/inputs/slt_text_field.dart';
+import 'package:go_router/go_router.dart';
+import 'package:spaced_learning_app/core/constants/app_constants.dart';
+import 'package:spaced_learning_app/core/theme/app_dimens.dart';
+import 'package:spaced_learning_app/presentation/screens/auth/register_screen.dart';
+import 'package:spaced_learning_app/presentation/viewmodels/auth_viewmodel.dart';
 
-import '../../../core/services/slt_ui_notifier_service.dart';
-import '../../../core/theme/app_dimens.dart';
-import '../../viewmodels/auth/login_view_model.dart';
-
-/// Login screen
-class LoginScreen extends ConsumerStatefulWidget {
-  const LoginScreen({Key? key}) : super(key: key);
+class LoginScreen extends ConsumerWidget {
+  const LoginScreen({super.key});
 
   @override
-  ConsumerState<LoginScreen> createState() => _LoginScreenState();
-}
-
-class _LoginScreenState extends ConsumerState<LoginScreen> {
-  final _formKey = GlobalKey<FormState>();
-  final _emailController = TextEditingController();
-  final _passwordController = TextEditingController();
-
-  final _uiNotifier = DependencyInjection.locator<UiNotifierService>();
-
-  @override
-  void dispose() {
-    _emailController.dispose();
-    _passwordController.dispose();
-    super.dispose();
-  }
-
-  void _login() async {
-    final loginViewModel = ref.read(loginViewModelProvider.notifier);
-
-    // Validate form
-    if (!_formKey.currentState!.validate()) {
-      return;
-    }
-
-    // Hide keyboard
-    FocusScope.of(context).unfocus();
-
-    final email = _emailController.text.trim();
-    final password = _passwordController.text;
-
-    // Store whether this widget is mounted before the async gap
-    bool isSuccessful = await loginViewModel.login(email, password);
-
-    // Check if the widget is still mounted after the async operation
-    if (!mounted) return;
-
-    if (isSuccessful) {
-      _uiNotifier.showSuccessSnackBar(
-        context,
-        'Login successful',
-      );
-
-      // Navigate to home screen
-      AppRoutes.navigateReplacementTo(context, AppRoutes.home);
-    }
-  }
-
-  @override
-  Widget build(BuildContext context) {
-    final loginState = ref.watch(loginViewModelProvider);
+  Widget build(BuildContext context, WidgetRef ref) {
     final theme = Theme.of(context);
+    final authState = ref.watch(authStateProvider);
+    final authError = ref.watch(authErrorProvider);
+    final formState = ref.watch(loginFormStateProvider);
+    final formNotifier = ref.read(loginFormStateProvider.notifier);
+
+    // Controllers for text fields - populated with current state values
+    final usernameOrEmailController = TextEditingController(
+      text: formState.usernameOrEmail,
+    );
+    final passwordController = TextEditingController(text: formState.password);
+
+    // Handle login submission
+    Future<void> handleLogin() async {
+      final success = await formNotifier.submitLogin();
+      if (success && context.mounted) {
+        GoRouter.of(context).go('/');
+      }
+    }
+
+    if (authState.isLoading) {
+      return SlLoadingStateWidget.fullScreen(message: 'Logging in...');
+    }
 
     return Scaffold(
-      appBar: const SltAppBar(
-        title: 'Login',
-        showBackButton: false,
-      ),
-      body: GestureDetector(
-        onTap: () => FocusScope.of(context).unfocus(),
-        child: SingleChildScrollView(
-          padding: AppDimens.screenPadding,
-          child: Form(
-            key: _formKey,
+      backgroundColor: theme.colorScheme.surface,
+      body: SafeArea(
+        child: Center(
+          child: SingleChildScrollView(
+            padding: const EdgeInsets.all(AppDimens.paddingXXL),
             child: Column(
+              mainAxisAlignment: MainAxisAlignment.center,
               crossAxisAlignment: CrossAxisAlignment.stretch,
               children: [
-                // Logo and title
-                Center(
-                  child: Column(
-                    children: [
-                      const SizedBox(height: 40),
-                      Image.asset(
-                        AppAssets.logoLightMode,
-                        height: 80,
-                        width: 80,
-                      ),
-                      const SizedBox(height: 24),
-                      Text(
-                        'Welcome Back',
-                        style: theme.textTheme.headlineMedium,
-                      ),
-                      const SizedBox(height: 8),
-                      Text(
-                        'Sign in to continue',
-                        style: theme.textTheme.bodyLarge?.copyWith(
-                          color: theme.colorScheme.onSurface.withOpacity(0.7),
-                        ),
-                      ),
-                    ],
-                  ),
+                _buildHeader(theme),
+                if (authError != null) _buildErrorWidget(authError, theme, ref),
+                const SizedBox(height: AppDimens.spaceL),
+                _buildFormFields(
+                  theme,
+                  formState,
+                  formNotifier,
+                  usernameOrEmailController,
+                  passwordController,
+                  handleLogin,
                 ),
-
-                const SizedBox(height: 40),
-
-                // Error message
-                if (loginState.errorMessage != null) ...[
-                  Container(
-                    padding: const EdgeInsets.all(16),
-                    decoration: BoxDecoration(
-                      color: theme.colorScheme.error.withOpacity(0.1),
-                      borderRadius: BorderRadius.circular(8),
-                      border: Border.all(
-                        color: theme.colorScheme.error.withOpacity(0.5),
-                      ),
-                    ),
-                    child: Text(
-                      loginState.errorMessage!,
-                      style: theme.textTheme.bodyMedium?.copyWith(
-                        color: theme.colorScheme.error,
-                      ),
-                    ),
-                  ),
-                  const SizedBox(height: 24),
-                ],
-
-                // Email field
-                SltTextField(
-                  controller: _emailController,
-                  labelText: 'Email',
-                  hintText: 'Enter your email',
-                  prefixIcon: Icons.email_outlined,
-                  keyboardType: TextInputType.emailAddress,
-                  textInputAction: TextInputAction.next,
-                  validator:
-                      ref.read(loginViewModelProvider.notifier).validateEmail,
-                  onChanged: (_) =>
-                      ref.read(loginViewModelProvider.notifier).clearError(),
-                  enabled: !loginState.isLoading,
-                ),
-
-                const SizedBox(height: 16),
-
-                // Password field
-                SltPasswordTextField(
-                  controller: _passwordController,
-                  labelText: 'Password',
-                  hintText: 'Enter your password',
-                  validator: ref
-                      .read(loginViewModelProvider.notifier)
-                      .validatePassword,
-                  onChanged: (_) =>
-                      ref.read(loginViewModelProvider.notifier).clearError(),
-                  textInputAction: TextInputAction.done,
-                  enabled: !loginState.isLoading,
-                  initiallyVisible: loginState.isPasswordVisible,
-                  onSubmitted: (_) => _login(),
-                ),
-
-                const SizedBox(height: 8),
-
-                // Forgot password
-                Align(
-                  alignment: Alignment.centerRight,
-                  child: TextButton(
-                    onPressed: loginState.isLoading
-                        ? null
-                        : () {
-                            // Navigate to forgot password screen
-                          },
-                    child: const Text('Forgot Password?'),
-                  ),
-                ),
-
-                const SizedBox(height: 24),
-
-                // Login button
-                SltPrimaryButton(
-                  text: 'Login',
-                  onPressed: loginState.isLoading ? null : _login,
-                  expandToFillWidth: true,
-                  isLoading: loginState.isLoading,
-                  height: 50,
-                ),
-
-                const SizedBox(height: 24),
-
-                // Register option
-                Row(
-                  mainAxisAlignment: MainAxisAlignment.center,
-                  children: [
-                    Text(
-                      'Don\'t have an account?',
-                      style: theme.textTheme.bodyMedium,
-                    ),
-                    TextButton(
-                      onPressed: loginState.isLoading
-                          ? null
-                          : () {
-                              // Navigate to register screen
-                              AppRoutes.navigateTo(context, AppRoutes.register);
-                            },
-                      child: const Text('Register'),
-                    ),
-                  ],
-                ),
-
-                const SizedBox(height: 16),
-
-                // Social login options
-                if (!loginState.isLoading) ...[
-                  const Center(
-                    child: Text('- OR -'),
-                  ),
-                  const SizedBox(height: 16),
-                  Row(
-                    mainAxisAlignment: MainAxisAlignment.center,
-                    children: [
-                      // Google login
-                      _buildSocialLoginButton(
-                        onPressed: () {
-                          // Implement Google login
-                        },
-                        icon: Icons.g_mobiledata,
-                        backgroundColor: Colors.white,
-                        borderColor: Colors.grey.shade300,
-                        iconColor: Colors.red,
-                      ),
-
-                      const SizedBox(width: 16),
-
-                      // Facebook login
-                      _buildSocialLoginButton(
-                        onPressed: () {
-                          // Implement Facebook login
-                        },
-                        icon: Icons.facebook,
-                        backgroundColor: Colors.blue.shade800,
-                        iconColor: Colors.white,
-                      ),
-
-                      const SizedBox(width: 16),
-
-                      // Apple login
-                      _buildSocialLoginButton(
-                        onPressed: () {
-                          // Implement Apple login
-                        },
-                        icon: Icons.apple,
-                        backgroundColor: Colors.black,
-                        iconColor: Colors.white,
-                      ),
-                    ],
-                  ),
-                ],
-
-                const SizedBox(height: 40),
+                const SizedBox(height: AppDimens.spaceXL),
+                _buildActions(context, theme, handleLogin),
               ],
             ),
           ),
@@ -272,33 +67,117 @@ class _LoginScreenState extends ConsumerState<LoginScreen> {
     );
   }
 
-  Widget _buildSocialLoginButton({
-    required VoidCallback onPressed,
-    required IconData icon,
-    required Color backgroundColor,
-    Color iconColor = Colors.white,
-    Color? borderColor,
-  }) {
-    return Material(
-      color: backgroundColor,
-      shape: CircleBorder(
-        side: borderColor != null
-            ? BorderSide(color: borderColor)
-            : BorderSide.none,
-      ),
-      elevation: 1,
-      child: InkWell(
-        onTap: onPressed,
-        customBorder: const CircleBorder(),
-        child: Padding(
-          padding: const EdgeInsets.all(12.0),
-          child: Icon(
-            icon,
-            color: iconColor,
-            size: 24,
+  Widget _buildHeader(ThemeData theme) {
+    return Column(
+      children: [
+        Text(
+          AppConstants.appName,
+          style: theme.textTheme.headlineMedium?.copyWith(
+            fontWeight: FontWeight.bold,
+            color: theme.colorScheme.primary,
           ),
+          textAlign: TextAlign.center,
         ),
+        const SizedBox(height: AppDimens.spaceS),
+        Text(
+          'Login to your account',
+          style: theme.textTheme.titleMedium?.copyWith(
+            color: theme.colorScheme.onSurfaceVariant,
+          ),
+          textAlign: TextAlign.center,
+        ),
+      ],
+    );
+  }
+
+  Widget _buildErrorWidget(
+    String errorMessage,
+    ThemeData theme,
+    WidgetRef ref,
+  ) {
+    return Padding(
+      padding: const EdgeInsets.only(top: AppDimens.paddingL),
+      child: SlErrorStateWidget(
+        title: 'Login Failed',
+        message: errorMessage,
+        compact: true,
+        onRetry: () => ref.read(authErrorProvider.notifier).clearError(),
+        icon: Icons.login_outlined,
+        accentColor: theme.colorScheme.error,
       ),
+    );
+  }
+
+  Widget _buildFormFields(
+    ThemeData theme,
+    LoginFormModel formState,
+    LoginFormState formNotifier,
+    TextEditingController usernameOrEmailController,
+    TextEditingController passwordController,
+    VoidCallback handleLogin,
+  ) {
+    return Column(
+      crossAxisAlignment: CrossAxisAlignment.stretch,
+      children: [
+        SLTextField(
+          label: 'Username or Email',
+          hint: 'Enter your username or email',
+          controller: usernameOrEmailController,
+          keyboardType: TextInputType.text,
+          errorText: formState.usernameOrEmailError,
+          prefixIcon: Icons.person,
+          onChanged: (value) => formNotifier.updateUsernameOrEmail(value),
+          textInputAction: TextInputAction.next,
+        ),
+        const SizedBox(height: AppDimens.spaceL),
+        SLPasswordField(
+          label: 'Password',
+          hint: 'Enter your password',
+          controller: passwordController,
+          errorText: formState.passwordError,
+          prefixIconData: Icons.lock_outline,
+          onChanged: (value) => formNotifier.updatePassword(value),
+          textInputAction: TextInputAction.done,
+          onEditingComplete: handleLogin,
+        ),
+      ],
+    );
+  }
+
+  Widget _buildActions(
+    BuildContext context,
+    ThemeData theme,
+    VoidCallback handleLogin,
+  ) {
+    return Column(
+      crossAxisAlignment: CrossAxisAlignment.stretch,
+      children: [
+        SltPrimaryButton(
+          text: 'Login',
+          onPressed: handleLogin,
+          isFullWidth: true,
+          size: SltButtonSize.large,
+        ),
+        const SizedBox(height: AppDimens.spaceL),
+        Row(
+          mainAxisAlignment: MainAxisAlignment.center,
+          children: [
+            Text(
+              "Don't have an account? ",
+              style: theme.textTheme.bodyMedium?.copyWith(
+                color: theme.colorScheme.onSurfaceVariant,
+              ),
+            ),
+            SltTextButton(
+              text: 'Register',
+              onPressed: () => Navigator.of(
+                context,
+              ).push(MaterialPageRoute(builder: (_) => const RegisterScreen())),
+              foregroundColor: theme.colorScheme.primary,
+            ),
+          ],
+        ),
+      ],
     );
   }
 }
