@@ -1,1 +1,48 @@
-// Outlined button placeholder 
+// lib/presentation/widgets/common/button/slt_outlined_button.dart
+import 'package:flutter/material.dart';
+import 'package:flutter_riverpod/flutter_riverpod.dart';
+import 'package:spaced_learning_app/presentation/widgets/buttons/slt_button_base.dart';
+
+class SltOutlinedButton extends ConsumerWidget {
+  final String text;
+  final VoidCallback? onPressed;
+  final IconData? prefixIcon;
+  final IconData? suffixIcon;
+  final String? loadingId;
+  final bool isFullWidth;
+  final SlttButtonSize size;
+  final Color? foregroundColor;
+  final BorderSide? borderSide;
+  final double? borderRadius;
+
+  const SltOutlinedButton({
+    super.key,
+    required this.text,
+    this.onPressed,
+    this.prefixIcon,
+    this.suffixIcon,
+    this.loadingId,
+    this.isFullWidth = false,
+    this.size = SlttButtonSize.medium,
+    this.foregroundColor,
+    this.borderSide,
+    this.borderRadius,
+  });
+
+  @override
+  Widget build(BuildContext context, WidgetRef ref) {
+    return SlttButtonBase(
+      text: text,
+      onPressed: onPressed,
+      prefixIcon: prefixIcon,
+      suffixIcon: suffixIcon,
+      loadingId: loadingId,
+      isFullWidth: isFullWidth,
+      size: size,
+      variant: SlttButtonVariant.outlined,
+      foregroundColor: foregroundColor,
+      borderSide: borderSide,
+      borderRadius: borderRadius,
+    );
+  }
+}
